sigsegv

+++ kernel/src/userspace.rs
@@ -1,7 +1,7 @@
-use self::process::{Process, ProcessId, PROCESSES};
+use self::process::{Process, PROCESSES};
 use crate::halt;
 
-pub const STACK_TOP: usize = 0x7ffffffff000; // Top of lower half but page aligned
+pub const STACK_TOP: u64 = 0x7ffffffff000; // Top of lower half but page aligned
 pub const INITIAL_STACK_SIZE_PAGES: usize = 16; // 64kib stack
 
 pub mod process;
